refactor: consolidate parallel and blocking devices
