+++ firmware/development/kampela-system/src/lib.rs
@@ -15,7 +15,6 @@ pub mod init;
 mod peripherals;
 pub mod devices;
 pub mod draw;
-pub mod parallel_devices;
 
 use efm32pg23_fix::{CorePeripherals, Peripherals};
